feat: use stream to reduce memory allocation during encode.
