+++ ua/buffer.go
@@ -9,8 +9,6 @@ import (
 	"io"
 	"math"
 	"time"
-
-	"github.com/gopcua/opcua/errors"
 )
 
 const (
@@ -202,125 +200,4 @@ func (b *Buffer) ReadN(n int) []byte {
 	}
 	b.pos += n
 	return d[:n]
-}
-
-func (b *Buffer) WriteBool(v bool) {
-	if v {
-		b.WriteUint8(1)
-	} else {
-		b.WriteUint8(0)
-	}
-}
-
-func (b *Buffer) WriteByte(n byte) {
-	b.buf = append(b.buf, n)
-}
-
-func (b *Buffer) WriteInt8(n int8) {
-	b.buf = append(b.buf, byte(n))
-}
-
-func (b *Buffer) WriteUint8(n uint8) {
-	b.buf = append(b.buf, byte(n))
-}
-
-func (b *Buffer) WriteInt16(n int16) {
-	b.WriteUint16(uint16(n))
-}
-
-func (b *Buffer) WriteUint16(n uint16) {
-	d := make([]byte, 2)
-	binary.LittleEndian.PutUint16(d, n)
-	b.Write(d)
-}
-
-func (b *Buffer) WriteInt32(n int32) {
-	b.WriteUint32(uint32(n))
-}
-
-func (b *Buffer) WriteUint32(n uint32) {
-	d := make([]byte, 4)
-	binary.LittleEndian.PutUint32(d, n)
-	b.Write(d)
-}
-
-func (b *Buffer) WriteInt64(n int64) {
-	b.WriteUint64(uint64(n))
-}
-
-func (b *Buffer) WriteUint64(n uint64) {
-	d := make([]byte, 8)
-	binary.LittleEndian.PutUint64(d, n)
-	b.Write(d)
-}
-
-func (b *Buffer) WriteFloat32(n float32) {
-	if math.IsNaN(float64(n)) {
-		b.WriteUint32(f32qnan)
-	} else {
-		b.WriteUint32(math.Float32bits(n))
-	}
-}
-
-func (b *Buffer) WriteFloat64(n float64) {
-	if math.IsNaN(n) {
-		b.WriteUint64(f64qnan)
-	} else {
-		b.WriteUint64(math.Float64bits(n))
-	}
-}
-
-func (b *Buffer) WriteString(s string) {
-	if s == "" {
-		b.WriteUint32(null)
-		return
-	}
-	b.WriteByteString([]byte(s))
-}
-
-func (b *Buffer) WriteByteString(d []byte) {
-	if b.err != nil {
-		return
-	}
-	if len(d) > math.MaxInt32 {
-		b.err = errors.Errorf("value too large")
-		return
-	}
-	if d == nil {
-		b.WriteUint32(null)
-		return
-	}
-	b.WriteUint32(uint32(len(d)))
-	b.Write(d)
-}
-
-func (b *Buffer) WriteStruct(w interface{}) {
-	if b.err != nil {
-		return
-	}
-	var d []byte
-	switch x := w.(type) {
-	case BinaryEncoder:
-		d, b.err = x.Encode()
-	default:
-		d, b.err = Encode(w)
-	}
-	b.Write(d)
-}
-
-func (b *Buffer) WriteTime(v time.Time) {
-	d := make([]byte, 8)
-	if !v.IsZero() {
-		// encode time in "100 nanosecond intervals since January 1, 1601"
-		ts := uint64(v.UTC().UnixNano()/100 + 116444736000000000)
-		binary.LittleEndian.PutUint64(d, ts)
-	}
-	b.Write(d)
-}
-
-func (b *Buffer) Write(d []byte) {
-	if b.err != nil {
-		return
-	}
-	b.buf = append(b.buf, d...)
 }
